# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Sistem Pelaporan Prestasi Mahasiswa API
  version: 1.0.0
  description: |
    REST API untuk sistem pelaporan prestasi mahasiswa dengan dukungan:
    - Role-Based Access Control (RBAC)
    - JWT Authentication
    - Achievement management dengan field dinamis
    - Verifikasi prestasi oleh dosen wali
    - Reporting dan analytics
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user dengan username dan password, mendapatkan JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Username atau password salah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh token
      description: Refresh JWT token menggunakan refresh token
      operationId: refreshToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token berhasil di-refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Token tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      description: Mendapatkan informasi user dari JWT token
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User info berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          user_id:
                            type: string
                            format: uuid
                          username:
                            type: string
                          role:
                            type: string
                          permissions:
                            type: array
                            items:
                              type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Mendapatkan profile lengkap user yang sedang login
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Logout user (saat ini hanya return success, token masih valid sampai expired)
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          user_id:
                            type: string
                            format: uuid
                          username:
                            type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token dari login endpoint
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: Pesan error dalam bahasa Indonesia
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: mahasiswa123
        password:
          type: string
          format: password
          example: SecurePass123!
    LoginResponse:
      type: object
      properties:
        error:
          type: boolean
          example: false
        data:
          type: object
          properties:
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIs...
            refreshToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIs...
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                username:
                  type: string
                fullName:
                  type: string
                role:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
    ProfileResponse:
      type: object
      properties:
        error:
          type: boolean
          example: false
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            fullName:
              type: string
            role:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
                description:
                  type: string
            permissions:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  resource:
                    type: string
                  action:
                    type: string
                  description:
                    type: string
            isActive:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...